+++ packages/styleguide/stories/Atoms/ProgressBar/ProgressBar.stories.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 
 import { ProgressBar, LayoutGrid, Column } from '@codecademy/gamut/src';
-import { colors } from '@codecademy/gamut-styles/utils/variables';
 
 import styles from './styles.module.scss';
 import {
@@ -11,26 +10,20 @@ import {
   StoryDescription,
   decoratedStory,
 } from '../../Templating';
+import { number } from '@storybook/addon-knobs';
 
 export default decoratedStories('Atoms', ProgressBar);
 
 const bars = [
   {
     large: false,
-    style: {
-      backgroundColor: colors.blue[100],
-      barColor: colors.blue[700],
-    },
+    theme: 'blue',
   },
   {
     large: true,
-    style: {
-      backgroundColor: colors.gray[100],
-      barColor: colors.yellow[500],
-      fontColor: colors.black,
-    },
+    theme: 'yellow',
   },
-];
+] as const;
 
 export const progressBar = decoratedStory(() => (
   <StoryTemplate status={StoryStatus.InProgress}>
@@ -45,12 +38,30 @@ export const progressBar = decoratedStory(() => (
     </StoryDescription>
     <LayoutGrid className={styles.progressBarGrid} columnGap="sm" rowGap="sm">
       {[0, 25, 50, 75, 100].map(percent =>
-        bars.map(({ large, style }) => (
-          <Column key={[percent, large, style].join()} size={6}>
-            <ProgressBar large={large} percent={percent} style={style} />
+        bars.map(({ large, theme }) => (
+          <Column key={[percent, large, theme].join()} size={6}>
+            <ProgressBar large={large} percent={percent} theme={theme} />
           </Column>
         ))
       )}
     </LayoutGrid>
   </StoryTemplate>
 ));
+
+export const progressBarMinimumPercent = decoratedStory(
+  'Minimum Percentage',
+  () => (
+    <StoryTemplate status={StoryStatus.InProgress}>
+      <StoryDescription>
+        Some bars (generally small ones) should display at least a little bit of
+        progress, even if the technical progress number is zero. You can use the
+        <code>minimumPercent</code> prop for a minimum visual width percentage.
+      </StoryDescription>
+      <ProgressBar
+        minimumPercent={number('Minimum Percent', 5)}
+        percent={number('Percent', 0)}
+        theme="blue"
+      />
+    </StoryTemplate>
+  )
+);